Preserve session info when removed from schedule (#302)

+++ funnel/views/session.py
@@ -17,19 +17,19 @@ def rooms_list(project):
 
 
 def session_form(project, proposal=None, session=None):
+    # Look for any existing unscheduled session
+    if proposal and not session:
+        session = Session.for_proposal(proposal)
+
     if session:
         form = SessionForm(obj=session, model=Session)
     else:
         form = SessionForm()
+
     form.venue_room_id.choices = rooms_list(project)
     if request.method == 'GET':
         if not (session or proposal):
             form.is_break.data = True
-        if proposal:
-            form.description.data = proposal.description
-            form.speaker_bio.data = proposal.bio
-            form.speaker.data = proposal.owner.fullname
-            form.title.data = proposal.title
         return render_template('session_form.html.jinja2', form=form, formid='session_form')
     if form.validate_on_submit():
         new = False
@@ -41,8 +41,6 @@ def session_form(project, proposal=None, session=None):
         form.populate_obj(session)
         if new:
             session.parent = project
-            session.make_id()  # FIXME: This should not be required
-            session.make_name()
             session = failsafe_add(db.session, session, project_id=project.id, url_id=session.url_id)
         db.session.commit()
         data = dict(
@@ -111,6 +109,9 @@ def session_edit(profile, project, session):
     permission='edit-session')
 def session_delete(profile, project, session):
     modal_url = session.proposal.url_for('schedule') if session.proposal else None
-    db.session.delete(session)
+    if not session.proposal:
+        db.session.delete(session)
+    else:
+        session.make_unscheduled()
     db.session.commit()
     return jsonify(status=True, modal_url=modal_url)
